adapt to respect linter syntax
